+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Replaced the package for custom svg icons with the MatIconRegistry
+- Replaced the lodash package with custom helpers
 
 
 ## [0.37.0] - 2025-01-17